+++ src/java/ring/adapter/jdk/KW.java
@@ -8,7 +8,6 @@ public class KW {
     public static final Keyword protocol = Keyword.intern("protocol");
     public static final Keyword remote_addr = Keyword.intern("remote-addr");
     public static final Keyword request_method = Keyword.intern("request-method");
-    @SuppressWarnings("unused")
     public static final Keyword scheme = Keyword.intern("scheme");
     public static final Keyword headers = Keyword.intern("headers");
     public static final Keyword server_name = Keyword.intern("server-name");
@@ -21,6 +20,13 @@ public class KW {
     public static final Keyword delete = Keyword.intern("delete");
     public static final Keyword patch = Keyword.intern("patch");
     public static final Keyword options = Keyword.intern("options");
+    public static final Keyword http = Keyword.intern("http");
+    @SuppressWarnings("unused")
+    public static final Keyword https = Keyword.intern("https");
+    @SuppressWarnings("unused")
+    public static final Keyword ws = Keyword.intern("ws");
+    @SuppressWarnings("unused")
+    public static final Keyword wss = Keyword.intern("wss");
 
     public static String KWtoString(final Keyword kw) {
         return kw.toString().substring(1);