+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 
 ## 0.1.1-SNAPSHOT
 
-- ?
+- fix scheme request field
+- fix uri with a query-string case
+- add tests
+- add todos
 
 ## 0.1.0